% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmvas.R
\name{lmvas}
\alias{lmvas}
\title{Bayesian linear model for EQ-5D using the score and visual analogue scale (VAS)}
\usage{
lmvas(vas_name, eq_name, independent_vars = NULL, data)
}
\arguments{
\item{vas_name}{variable containing the VAS (character)}

\item{eq_name}{variable containing the EQ-5D score (character)}

\item{independent_vars}{vector of independent variables (vector)}

\item{data}{data frame containing the independent variables and EQ-5D outcomes}
}
\value{
inla
}
\description{
Runs a Bayesian multivariate linear regression model using INLA.
The two dependent variables are the EQ-5D score and VAS which are assumed to be correlated.
}
\details{
inputs:
}
\examples{
model = res = lmvas(vas_name='EQ5D_VAS', eq_name='EQ5D', independent_vars =
c('age','sex'), data=arthritis)
summary(model)
}
